+++ CHANGES.rst
@@ -10,6 +10,7 @@ CHANGELOG for LaunchKey Python SDK
 * Added auth_methods and auth_policy attributes to the AuthorizationResponse object
 * Added handle_webhook as well as DirectoryUserDeviceLinkCompletionWebhookPackage into the DirectoryClient
 * Updated the OrganizationClient update_directory() method and Directory object to include a webhook_url kwarg / attribute
+* Added Webhook example app
 
 3.4.0
 -----